Integrate the remote main update before publishing the DirectWrite CI fix.

Testing: meson test -C build --print-errorlogs

+++ src/hb-raster-draw.cc
@@ -450,7 +450,7 @@ hb_raster_draw_get_edge_work (hb_raster_draw_t *draw, unsigned max_rows)
 {
   int64_t work = 0;
   for (const auto &e : draw->edges)
-    work += 1 + hb_min ((int64_t) (e.yH - e.yL) >> HB_RASTER_PIXEL_BITS,
+    work += 1 + hb_min (((int64_t) e.yH - (int64_t) e.yL) >> HB_RASTER_PIXEL_BITS,
 			(int64_t) max_rows);
   return work;
 }
